requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch
torchvision
huggingface-hub
Pillow
timm
transformers